tags/v0.5.0/marmot/requirements_web.txt
Django==1.8.13
MySQL-python==1.2.5
redis==2.10.5
zeroc-ice==3.6.2.1